+++ src/Laplace.Core/Services/ArchiveFormatDetector.cs
@@ -72,7 +72,7 @@ public static class ArchiveFormatDetector
         throw new NotSupportedException($"Unsupported output archive format '{extension}'. Supported write formats: {SupportedWriteFormats}");
     }
 
-    public static string SupportedReadFormats => ".lpc, .zip, .7z, .rar, .tar, .tar.gz/.tgz, .tar.bz2/.tbz2, .tar.xz/.txz, .gz, .bz2, .xz, .zst, .lzip";
+    public static string SupportedReadFormats => ".lpc, .zip, .7z, .rar, .iso, .tar, .tar.gz/.tgz, .tar.bz2/.tbz2, .tar.xz/.txz, .gz, .bz2, .xz, .zst, .lzip";
     public static string SupportedWriteFormats => ".lpc, .zip, .7z, .rar";
 
     private static bool LooksLikeLpc(string archivePath)